requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs
aiohttp
aiosignal
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
antlr4-python3-runtime==4.9.3
asttokens
astunparse==1.6.3
async-timeout
attrs
audioread==3.0.1
av==13.1.0
axial_positional_embedding==0.2.1
backcall
blobfile==3.0.0
Bottleneck
Brotli
certifi
cffi==1.17.1
charset-normalizer
click==8.1.8
cmake==3.30.3
colorama==0.4.6
comm
contourpy==1.3.0
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
datasets
debugpy
decorator==4.4.2
Deprecated==1.2.14
dill
ecg-plot==0.2.8
einops==0.6.0
entrypoints
executing
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist
fsspec
ftfy==6.3.0
fvcore==0.1.5.post20221221
gast==0.6.0
gmpy2
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
huggingface-hub==0.26.0
hydra-core==1.3.2
idna==3.10
imagecodecs
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
ipykernel
ipython
jedi
Jinja2
jmespath==0.10.0
joblib==1.4.2
jupyter-client
jupyter_core
keras==3.6.0
kiwisolver==1.4.7
lazy_loader
libclang==18.1.1
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities
lit==18.1.8
llvmlite==0.43.0
local-attention==1.9.14
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe
mat73==0.65
matplotlib==3.7.0
matplotlib-inline
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft
mkl_random
ml-dtypes==0.4.1
mmcv==2.2.0
mmcv-full==1.7.2
mmengine==0.7.4
model-index==0.1.11
moviepy @ git+https://github.com/Zulko/moviepy.git@0f6f6d4d9b96c69d109975549fe6293931f1a19d
mpmath
msgpack==1.1.0
multidict
multiprocess
namex==0.0.8
nest_asyncio
networkx
neurokit2==0.2.10
numba==0.60.0
numexpr
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
opt_einsum==3.4.0
optree==0.13.0
ordered-set==4.1.0
oss2==2.17.0
packaging
pandas==1.5.3
parameterized==0.9.0
parso
patool==1.12
pexpect
pickleshare
pillow
platformdirs
pooch==1.8.2
portalocker==2.10.1
product-key-memory==0.2.2
proglog==0.1.10
prompt_toolkit
protobuf==5.28.2
psutil
ptyprocess
pure_eval
pyarrow
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@2723c7fa4f2ead16ae1ce3d8afe977724bb8f87f
Pygments
pyparsing==3.1.4
PySocks
python-dateutil
pytorch-lightning==1.8.4
pytorchvideo==0.1.5
pytz==2023.4
PyWavelets==1.6.0
PyYAML
pyzmq
reformer-pytorch==1.4.4
regex
requests==2.28.2
rich==13.4.2
safetensors
scikit-image
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.2
sentencepiece==0.2.0
six
sktime==0.16.1
soundfile==0.12.1
soxr==0.5.0.post1
stack-data
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile
tiktoken==0.8.0
timm==1.0.11
tokenizers
tomli==2.2.1
torch==2.0.1+cu118
torchao==0.6.1
torchaudio==2.0.2+cu118
torchmetrics
torchtune==0.4.0
torchvision==0.15.2+cu118
tornado
tqdm==4.65.2
traitlets
transformers
triton==2.0.0
ts2vg==1.2.4
typing_extensions
tzdata
urllib3==1.26.20
wcwidth
Werkzeug==3.0.6
wfdb==4.1.2
wget==3.2
wrapt==1.16.0
xxhash
yacs==0.1.8
yapf==0.43.0
yarl
zipp==3.20.2

